:bug: Improved detection of terminating pods

+++ pkg/controller/chaospod/chaospod_controller.go
@@ -22,11 +22,6 @@ import (
 
 var log = logf.Log.WithName("controller_chaospod")
 
-/**
-* USER ACTION REQUIRED: This is a scaffold file intended for the user to modify with their own Controller
-* business logic.  Delete these comments after modifying this file.*
- */
-
 // Add creates a new ChaosPod Controller and adds it to the Manager. The Manager will set fields on the Controller
 // and Start it when the Manager is Started.
 func Add(mgr manager.Manager) error {
@@ -72,8 +67,6 @@ type ReconcileChaosPod struct {
 
 // Reconcile reads that state of the cluster for a ChaosPod object and makes changes based on the state read
 // and what is in the ChaosPod.Spec
-// TODO(user): Modify this Reconcile function to implement your Controller logic.  This example creates
-// a Pod as an example
 // Note:
 // The Controller will requeue the Request to be processed again if the returned error is non-nil or
 // Result.Requeue is true, otherwise upon completion it will remove the work from the queue.
@@ -136,14 +129,9 @@ func killPods(r *ReconcileChaosPod, chaosPod *chaosv1alpha1.ChaosPod, existingPo
 
 	reqLogger.Info("Searching for pods with prefix " + prefixToKill)
 	for _, pod := range existingPods {
-		if strings.HasPrefix(pod.Name, prefixToKill) {
+		isAlreadyBeeingTerminated := pod.GetDeletionTimestamp() != nil
+		if strings.HasPrefix(pod.Name, prefixToKill) && !isAlreadyBeeingTerminated {
 			podName := pod.Name
-			podUID := string(pod.UID)
-			alreadyKilledPods := chaosPod.Status.KilledPodNames
-			if _, ok := alreadyKilledPods[podUID]; ok {
-				reqLogger.Info("😥 Attempted to kill previously killed pod, assume its beeing terminated.", "Pod.Namespace", pod.Namespace, "Pod.Name", podName)
-				continue
-			}
 
 			reqLogger.Info("🎉 Yay! Found pod to kill!", "Pod.Namespace", pod.Namespace, "Pod.Name", podName)
 			err := r.client.Delete(context.TODO(), &pod)
@@ -151,7 +139,7 @@ func killPods(r *ReconcileChaosPod, chaosPod *chaosv1alpha1.ChaosPod, existingPo
 			if err != nil {
 				logDeletePodError(reqLogger, err, podName)
 			} else {
-				killedPodNames[podUID] = podName
+				killedPodNames[string(pod.UID)] = podName
 				reqLogger.Info("💀 Killed pod!", "Pod.Namespace", pod.Namespace, "Pod.Name", podName)
 			}
 		}
